+++ src/components/ButtonBack/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useHistory } from "react-router-dom";
 import IconBack from "../Icons/IconBack";
 
-const ButtonBack = () => {
+const ButtonBack = (): JSX.Element => {
     const history = useHistory();
 
     const handleClick = () => {
