+++ src/components/Form/types.ts
@@ -1,0 +1,8 @@
+import { FormEvent, ReactNode } from "react";
+
+export interface FormProps {
+    handleSubmit?: (event: FormEvent) => void;
+    children: ReactNode;
+    className?: string;
+    isContentCentered?: boolean;
+}
