+++ src/components/Header/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { StyledHeader } from "./styles";
+import { HeaderProps } from "./types";
+
+const Header: React.FC<HeaderProps> = ({ children, className }: HeaderProps) => {
+    return (
+        <StyledHeader className={className}>
+            <nav>{children}</nav>
+        </StyledHeader>
+    );
+};
+
+export default Header;
