+++ src/components/Header/styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const StyledHeader = styled.header`
+    height: 10vh;
+    display: flex;
+    align-items: center;
+    padding: 0 0.5rem;
+`;
