+++ src/components/Header/types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from "react";
+
+export interface HeaderProps {
+    children: ReactNode;
+    className?: string;
+}
